pomodoro_bot/requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0
numpy==1.26.3
python-dotenv==1.0.0
